pyproject.toml
[project]
name = "id-gen"
version = "0.1.0"
description = "諾瓦雷克斯帝國身份證生成器"
authors = [
    { name = "SamHacker", email = "xux510208@gmail.com" },
    { name = "諾瓦雷克斯帝國數位證物府" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.3.0",
    "python-barcode>=0.15.1",
    "svglib>=1.5.1",
    "colorlog>=6.7.0",
    "click>=8.2.1",
    "tabulate>=0.9.0",
    "termcolor>=3.1.0",
]

[dependency-groups]
dev = [
    "nuitka>=2.7.12",
]
